+++ crates/ic-query/src/nns/node/report/text/mod.rs
@@ -1,7 +1,9 @@
 mod info;
 mod list;
+#[cfg(feature = "host")]
 mod refresh;
 
 pub use info::nns_node_info_report_text;
 pub use list::{nns_node_list_report_text, nns_node_list_report_verbose_text};
+#[cfg(feature = "host")]
 pub use refresh::nns_node_refresh_report_text;